TGBot/tgBot/requirements.txt
python-telegram-bot==22.0
requests>=2.28.0
PyYAML>=6.0
pydantic>=1.10.0
